requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.9
django-ckeditor==5.4.0
django-colorfield==0.1.14
django-filter==1.1.0
django-hvad==1.8.0
django-js-asset==1.0.0
django-mptt==0.9.0
django-parler==1.9.2
django-redis==4.9.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.0.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pytz==2017.3
redis==2.10.6
requests==2.18.4
simplejson==3.13.2
solid-i18n==1.4.1
uritemplate==3.0.0
urllib3==1.22
yattag==1.10.0
